Refactored Joining Query
